backend/requirements.txt
# Framework Web
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0

# Base de Datos
mysql-connector-python==8.2.0
pymysql==1.1.0

# Machine Learning y Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Reconocimiento Facial
face-recognition>=1.3.0
cmake>=3.25.0
dlib>=19.24.0
pillow>=10.0.0

# IA y NLP
google-generativeai>=0.3.0

# Utilidades
python-dotenv>=1.0.0
bcrypt>=4.1.0
python-dateutil>=2.8.0
